Changes: move test helper
